pyproject.toml
[project]
name = "aio-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Hieu",email = "hieutm@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain-openai (>=0.3.1,<0.4.0)",
    "langchain-core (>=0.3.30,<0.4.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "tavily-python (>=0.5.0,<0.6.0)",
    "langgraph (>=0.2.64,<0.3.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "crewai (>=0.98.0,<0.99.0)",
    "langchain-google-community[gmail] (>=2.0.4,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pymongo (>=4.10.1,<5.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "langchain-mongodb (>=0.4.0,<0.5.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "docx2txt (>=0.8,<0.9)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
